irecognise-backend/requirements.txt
Flask
Flask-Bcrypt
python-dotenv
pymongo
imutils
sockets
opencv-python
imagezmq